requirements.txt
gspread
pandas
qrcode
Pillow
google-auth-oauthlib
google-api-python-client